server/requirements.txt
Django==1.5.3
OpenPGP==0.2.3
distribute==0.6.24
dj-database-url==0.2.2
dj-static==0.0.5
django-mailgun==0.2.2
django-registration==1.0
django-toolbelt==0.0.1
gunicorn==18.0
ipdb==0.7
ipython==1.1.0
psycopg2==2.5.1
pycrypto==2.6
requests==1.2.3
static==0.4
